Question 4:
D. After the Parrot object is instantiated, a specific action Talk class
can be added to the Parrot to make it speak dog, cat or rooster.

In fact, the rooster talk object used for the rooster animal was also
re-used for the parrot.

Additional talk action objects can be created and added to make the
Parrot object maintainable.